Refactor: sever runtime → testing circular dependency (#528)
* Initial plan

* Refactor: sever runtime → testing circular dependency

Agent-Logs-Url: https://github.com/SergeiGolos/wod-wiki/sessions/c03fde9e-02e8-44a1-bedf-14c071372fb9


* refactor: use import type for type-only imports in ITestableBlockConfig, IRuntimeOptions, RuntimeBuilder, TestableBlock

Agent-Logs-Url: https://github.com/SergeiGolos/wod-wiki/sessions/708bfa6b-a7ac-4200-bb95-7c08c2010eea


---------
