+++ src/runtime/compiler/RuntimeBuilder.ts
@@ -6,7 +6,7 @@ import { EventBus } from '../events/EventBus';
 import { JitCompiler } from './JitCompiler';
 import { WodScript } from '../parser/WodScript';
 import { IRuntimeOptions, DEFAULT_RUNTIME_OPTIONS } from '../contracts/IRuntimeOptions';
-import { TestableBlockConfig } from '../../testing/testable/TestableBlock';
+import type { TestableBlockConfig } from '../contracts/ITestableBlockConfig';
 
 /**
  * RuntimeBuilder provides a fluent API for constructing ScriptRuntime instances
