--- src/runtime/testing/index.ts
@@ -1,84 +1,0 @@
-/**
- * Runtime Testing Infrastructure
- * 
- * This module provides tools for testing IRuntimeBlock implementations
- * by wrapping them with testable wrappers that track method calls,
- * memory operations, and stack changes.
- * 
- * ## Custom Wrapper for Debug Tracking
- * 
- * Provide a custom wrapper via options to track wrapped blocks:
- * 
- * ```typescript
- * import { ScriptRuntime, RuntimeStackWrapper } from '@/runtime';
- * import { TestableBlock } from '@/testing';
- * 
- * const wrappedBlocks = new Map<string, TestableBlock>();
- * 
- * const wrapper: RuntimeStackWrapper = {
- *   wrap: (block) => {
- *     const wrapped = new TestableBlock(block, { testId: `debug-${block.key}` });
- *     wrappedBlocks.set(block.key.toString(), wrapped);
- *     return wrapped;
- *   },
- *   cleanup: (block) => wrappedBlocks.delete(block.key.toString()),
- * };
- * 
- * const runtime = new ScriptRuntime(script, compiler, { wrapper });
- * 
- * // Execute workout...
- * 
- * // Inspect wrapped blocks
- * for (const [key, block] of wrappedBlocks) {
- *   console.log(`${key}: mount called ${block.callCount('mount')} times`);
- * }
- * ```
- * 
- * ## Manual Testing
- * 
- *   nextMode: 'spy'
- * });
- * 
- * // Wrap runtime to track operations
- * const testRuntime = new TestableRuntime(realRuntime, {
- *   initialMemory: [{ type: 'metric:reps', ownerId: 'parent', value: 21 }]
- * });
- * 
- * // Take snapshots and diff
- * const before = testRuntime.snapshot('before');
- * testableBlock.mount(testRuntime);
- * const after = testRuntime.snapshot('after');
- * const diff = testRuntime.diff(before, after);
- * ```
- */
-
-// Core testing classes
-export {
-  TestableBlock,
-  createTestableBlock,
-  type TestableBlockConfig,
-  type InterceptMode,
-  type MethodCall,
-  type MemoryOperation,
-  type StackOperation
-} from '../../testing/testable/TestableBlock';
-
-export {
-  TestableRuntime,
-  createTestableRuntime,
-  type TestableRuntimeConfig,
-  type InitialMemoryEntry,
-  type InitialStackEntry,
-  type RuntimeSnapshot,
-  type SnapshotDiff
-} from '../../testing/testable/TestableRuntime';
-
-// Visualization components
-export {
-  SnapshotDiffViewer,
-  SnapshotDiffSummary,
-  ModifiedValuesViewer
-} from '../../testing/components';
-
-// Test setup actions
-export * from '../../testing/setup';
